requirements.txt
# Compatible pinned versions
gradio==4.44.1
gradio_client==1.3.0
huggingface_hub==0.23.0
pydantic==2.10.6

# Your dependencies
torch
torchvision
pandas
requests
Pillow
numpy
geopy
earthengine-api


# gradio==4.44.0
# huggingface_hub==0.23.0
# torch
# torchvision
# pandas

# requests
# Pillow

# numpy
# earthengine-api
# geopy
